libstdc++: Use [] for empty argument to automake macro
Does this look reasonable to push? There are lots of other uses of
AM_ENABLE_MULTILIB(, ..) in the tree, but I'll leave others to "fix"
those if needed.

-- >8 --

This seems to be needed to build with --enable-maintainer-mode but I
don't understand why it doesn't fail in the normal build.

libstdc++-v3/ChangeLog:

	* configure.ac: Use [] for empty argument to AM_ENABLE_MULTILIB
	macro.