middle-end: Fix dominator information with loop duplication PR114197
When adding the new_preheader to the cfg, only the new_preheader's dominator
information is updated. If one of the new basic block's children was part
of the original cfg and adding new_preheader to the cfg introduces another path
to that child, the child's dominator information will not be updated. This may
cause verify_dominator's assertion to fail.

Force recalculating dominators for all duplicated basic blocks and their
successors when updating new_preheader's dominator information.

	PR 114197

gcc/ChangeLog:

	* tree-vect-loop-manip.cc (slpeel_tree_duplicate_loop_to_edge_cfg):
	Recalculate dominator info when adding new_preheader to cfg

gcc/testsuite/ChangeLog:

	* gcc.dg/vect/pr114197.c: New test.

Signed-off-by: Edwin Lu <ewlu@rivosinc.com>