c++: lambda capturing structured bindings [PR85889]
Bootstrapped/regtested on x86_64-pc-linux-gnu, ok for 15?  (Or even trunk?)

-- >8 --
<https://wg21.link/p1381r1> clarifies that it's OK to capture structured
bindings.

[expr.prim.lambda.capture]/4 says "The identifier in a simple-capture shall
denote a local entity" and [basic.pre]/3: "An entity is a [...] structured
binding".

It doesn't appear that this was made a DR, so, strictly speaking, we
should have a -Wc++20-extensions warning, like clang++.

	PR c++/85889

gcc/cp/ChangeLog:

	* lambda.cc (add_capture): Add a pedwarn for capturing structured
	bindings.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp2a/decomp3.C: Use -Wno-c++20-extensions.
	* g++.dg/cpp1z/decomp60.C: New test.