c++: Don't set DECL_CONTEXT to nested template-template parameters [PR98881]
Bootstrapped and regtested on x86_64-pc-linux-gnu, OK for trunk?

-- >8 --

When streaming in a nested template-template parameter as in the
attached testcase, we end up reaching the containing template-template
parameter in 'tpl_parms_fini'. We should not set the DECL_CONTEXT to
this (nested) template-template parameter, as it should already be the
struct that the outer template-template parameter is declared on.

	PR c++/98881

gcc/cp/ChangeLog:

	* module.cc (trees_out::tpl_parms_fini): Clarify logic purely
	for checking purposes. Don't consider a template template
	parameter as the owning template.
	(trees_in::tpl_parms_fini): Don't consider a template template
	parameter as the owning template.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/tpl-tpl-parm-3_a.H: New test.
	* g++.dg/modules/tpl-tpl-parm-3_b.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>