c++/modules: Implement P2615 'Meaningful Exports' [PR107688]
Bootstrapped and regtested on x86_64-pc-linux-gnu. This should probably
wait for GCC 15 I suppose, but sending it in now in case there are any
comments.

-- >8 --

This clarifies which kinds of declarations may and may not be exported
in various contexts. The patch additionally fixes up some small issues
that were clarified by the paper.

Most of the changes are with regards to export-declarations, which are
applied for all standards modes that we support '-fmodules-ts' for.
However there are also a couple of changes made to linkage specifiers
('extern "C"'); I've applied these as since C++20, to line up with when
modules were actually introduced.

	PR c++/107688

gcc/cp/ChangeLog:

	* name-lookup.cc (push_namespace): Error when exporting
	namespace with internal linkage.
	* parser.h (struct cp_parser): Add new flag
	'in_unbraced_export_declaration_p'.
	* parser.cc (cp_debug_parser): Print the new flag.
	(cp_parser_new): Initialise the new flag.
	(cp_parser_module_export): Set the new flag.
	(cp_parser_class_specifier): Clear and restore the new flag.
	(cp_parser_import_declaration): Imports can now appear directly
	in a linkage specification.
	(cp_parser_declaration): Categorise declarations as "name" or
	"special"; error on the later in contexts where the former is
	required.
	(cp_parser_class_head): Error when exporting a partial
	specialisation.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/contracts-1_a.C: Avoid now-illegal syntax.
	* g++.dg/modules/contracts-2_a.C: Likewise.
	* g++.dg/modules/contracts-3_a.C: Likewise.
	* g++.dg/modules/contracts-4_a.C: Likewise.
	* g++.dg/modules/lang-1_c.C: Clarify now-legal syntax.
	* g++.dg/template/crash71.C: Update error messages.
	* g++.dg/cpp2a/linkage-spec1.C: New test.
	* g++.dg/modules/export-3.C: New test.
	* g++.dg/modules/export-4_a.C: New test.
	* g++.dg/modules/export-4_b.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>